requirements.txt
alabaster==0.7.16
bibtexparser==1.4.1
blinker==1.7.0
click==8.1.7
Deprecated==1.2.14
fake-useragent==1.4.0
Flask==3.0.2
free_proxy==1.1.1
gunicorn==21.2.0
imagesize==1.4.1
itsdangerous==2.1.2
lxml==5.1.0
outcome==1.3.0.post0
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.1
scholarly==1.7.11
selenium==4.18.1
sortedcontainers==2.4.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
trio==0.24.0
trio-websocket==0.11.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
